requirements.txt
# 基础数据处理依赖
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0

# 自然语言处理
nltk==3.8.1
spacy==3.6.1
textblob==0.17.1

# LLM集成 - 优先使用Hugging Face免费模型
transformers==4.30.2
# openai==0.27.8  # 不再需要，可以注释掉
langchain==0.0.257

# 深度学习框架 - 用于运行Hugging Face模型
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
accelerate==0.20.3  # 优化transformers模型推理速度

# 数据获取
praw==7.7.1  # Reddit API
python-dotenv==1.0.0  # 环境变量管理

# 进度条 - 用于课堂展示
tqdm==4.65.0  # 显示处理进度条，提升课堂展示体验

# 其他工具
jupyter==1.0.0
notebook==6.5.4
plotly==5.16.0
wordcloud==1.9.2